update to take continuous input
